feat(MAM-6): Storybook preview 설정 파일을 .tsx로 변경 및 전역 스타일 적용
